@@ -8,12 +8,13 @@
   },
   "dependencies": {
     "element-ui": "^2.15.14",
-    "vue": "^2.7.7"
+    "vue": "^3.4.38",
+    "@vue/compat": "^3.4.38"
   },
   "devDependencies": {
-    "@vitejs/plugin-legacy": "^2.0.0",
-    "@vitejs/plugin-vue2": "^1.1.2",
-    "terser": "^5.14.2",
-    "vite": "^3.0.2"
+    "@vitejs/plugin-vue": "^5.1.2",
+    "@vue/compiler-sfc": "^3.4.38",
+    "terser": "^5.31.6",
+    "vite": "^5.4.1"
   }
 }
