@@ -1,12 +1,13 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import Element from 'element-ui'
 
 import './assets/main.css'
 import 'element-ui/lib/theme-chalk/index.css';
 
-Vue.use(Element)
+// Element UI: npm i --legacy-peer-deps
 
-new Vue({
-  render: (h) => h(App)
-}).$mount('#app')
+const app = createApp(App)
+
+app.use(Element)
+app.mount('#app')
